[BROKEN][WIP] JUMO.Core에서 tick의 타입을 long에서 int로 변경

+++ JUMO.Core/IMusicalItem.cs
@@ -2,7 +2,7 @@
 {
     public interface IMusicalItem
     {
-        long Length { get; }
-        long Start { get; }
+        int Length { get; }
+        int Start { get; }
     }
 }
